SHM_ functions in its own file
